Add simple HTTP request logging
The logging format is hardcoded and does not do any timing of requests
etc. The logger is an http.Handler that then runs ServeHTTP on another
http.Handler in order to serve files.

The http.Handler 'chaining' technique is actually more general than
this, if different functionality or additional http.Handlers are wanted
to run in sequence.